requirement.txt
fastapi
uvicorn
pymongo
bcrypt
python-jose[cryptography]
apscheduler
email-validator
pydantic[email]
dnspython
